Add working features extractor, list of currency is to short (need to expand it), most common country in progress.

+++ ksr/src/main/java/com/tul/ksr/zad1/model/Features.java
@@ -12,7 +12,7 @@ public class Features {
     String secondCommonCurrency;
     String mostCommonCountry;
     String secondCommonCountry;
-    int avgWordLength;
+    double avgWordLength;
     int wordShortThanFive;
 
     public Features(int articleLength, int numberOfWords, String longestWord, int lengthOfLongestWord,
